requirements.txt
fastapi
uvicorn[standard]
supabase>=2.3.0
google-adk
pydantic>=2.0
httpx
python-dotenv
